Fix Link-Ack issues on init
Link ack issues on init were being caused by the DelayMs blocking the software flow.  I've moved the init commands to after the system initialization has completed as well.

Getting the second command to run is a bit of a hack, but it works and the delays are removed after the function successfully completes.